Reduced camerasInspector rerender
